fix: mobile expand
