+++ apps/dashboard/components/Layout/index.tsx
@@ -1,9 +1,20 @@
+import AllInboxRoundedIcon from '@mui/icons-material/AllInboxRounded';
+import ApiRoundedIcon from '@mui/icons-material/ApiRounded';
+import ArrowForwardIosIcon from '@mui/icons-material/ArrowForwardIos';
+import ArrowForwardRoundedIcon from '@mui/icons-material/ArrowForwardRounded';
+import ChatRoundedIcon from '@mui/icons-material/ChatRounded';
+import FeedRoundedIcon from '@mui/icons-material/FeedRounded';
+import InboxRoundedIcon from '@mui/icons-material/InboxRounded';
 import LinkedInIcon from '@mui/icons-material/LinkedIn';
-import MailRoundedIcon from '@mui/icons-material/MailRounded';
-import MenuIcon from '@mui/icons-material/Menu';
+import ManageAccountsRoundedIcon from '@mui/icons-material/ManageAccountsRounded';
+import RecentActorsIcon from '@mui/icons-material/RecentActors';
+import ShowChartIcon from '@mui/icons-material/ShowChart';
+import SmartToyRoundedIcon from '@mui/icons-material/SmartToyRounded';
+import StorageRoundedIcon from '@mui/icons-material/StorageRounded';
 import TwitterIcon from '@mui/icons-material/Twitter';
 import {
   Alert,
+  Badge,
   Button,
   Chip,
   IconButton,
@@ -12,26 +23,35 @@ import {
   useColorScheme,
 } from '@mui/joy';
 import Box from '@mui/joy/Box';
-import { SxProps } from '@mui/joy/styles/types';
+import { ColorPaletteProp, SxProps } from '@mui/joy/styles/types';
 import Typography from '@mui/joy/Typography';
+import getConfig from 'next/config';
+import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { useSession } from 'next-auth/react';
 import React from 'react';
+import toast from 'react-hot-toast';
+import { useSWRConfig } from 'swr';
+import useSWR from 'swr';
 
 import useModal from '@app/hooks/useModal';
+import { useNavbar } from '@app/hooks/useNavbar';
+import useProduct from '@app/hooks/useProduct';
+import { countUnread } from '@app/pages/api/logs/count-unread';
+import { getStatus } from '@app/pages/api/status';
 
 import { appUrl } from '@chaindesk/lib/config';
-import DarkModeToggle from '@chaindesk/ui/DarkModeToggle';
+import { fetcher } from '@chaindesk/lib/swr-fetcher';
+import { AppStatus, RouteNames } from '@chaindesk/lib/types';
+import { Prisma } from '@chaindesk/prisma';
 
-import Logo from '../Logo';
 import SEO from '../SEO';
 
-import Header from './Header';
+import ExpandedNavigation from './ExpandedNavigation';
 import Main from './Main';
 import Navigation from './Navigation';
 import Root from './Root';
 import SideDrawer from './SideDrawer';
-import SideNav from './SideNav';
 
 type Props = {
   children: React.ReactNode;
@@ -42,6 +62,25 @@ export default function Layout(props: Props) {
   const router = useRouter();
   const { mode, setMode } = useColorScheme();
   const { data: session, status } = useSession();
+  const { product } = useProduct();
+  const { open, setOpen } = useNavbar();
+
+  const countUnreadQuery = useSWR<Prisma.PromiseReturnType<typeof countUnread>>(
+    '/api/logs/count-unread',
+    fetcher,
+    {
+      refreshInterval: 60000,
+    }
+  );
+
+  const getStatusQuery = useSWR<Prisma.PromiseReturnType<typeof getStatus>>(
+    '/api/status',
+    fetcher,
+    {
+      refreshInterval: 60000,
+    }
+  );
+
   const [drawerOpen, setDrawerOpen] = React.useState(false);
   const [userMenuElement, setUserMenuElement] =
     React.useState<null | HTMLElement>(null);
@@ -64,6 +103,244 @@ export default function Layout(props: Props) {
     setUserMenuElement(null);
   };
 
+  const appLinks = React.useMemo(() => {
+    return [
+      ...(product === 'chaindesk'
+        ? [
+            {
+              label: 'Inbox',
+              route: RouteNames.LOGS,
+              icon: (
+                <Badge
+                  badgeContent={countUnreadQuery?.data}
+                  size="sm"
+                  color="danger"
+                  invisible={
+                    !countUnreadQuery?.data || countUnreadQuery?.data <= 0
+                  }
+                >
+                  <InboxRoundedIcon style={{ fontSize: '18px' }} />
+                </Badge>
+              ),
+              active: router.route === RouteNames.LOGS,
+              isNew: false,
+            },
+            {
+              label: 'Agents',
+              route: RouteNames.AGENTS,
+              icon: <SmartToyRoundedIcon style={{ fontSize: '18px' }} />,
+              active: router.route.startsWith(RouteNames.AGENTS),
+              isExperimental: false,
+              isNew: false,
+            },
+            {
+              label: 'Datastores',
+              route: RouteNames.DATASTORES,
+              icon: <StorageRoundedIcon style={{ fontSize: '18px' }} />,
+              active: router.route.startsWith(RouteNames.DATASTORES),
+              isNew: false,
+            },
+            {
+              label: 'Forms',
+              route: RouteNames.FORMS,
+              icon: <FeedRoundedIcon style={{ fontSize: '18px' }} />,
+              active: router.route.startsWith(RouteNames.FORMS),
+              isNew: false,
+              isExperimental: false,
+            },
+            {
+              label: 'Analytics',
+              route: RouteNames.ANALYTICS,
+              icon: <ShowChartIcon style={{ fontSize: '18px' }} />,
+              active: router.route.startsWith(RouteNames.ANALYTICS),
+              isNew: false,
+            },
+            {
+              label: 'Email Inboxes',
+              route: RouteNames.EMAIL_INBOXES,
+              icon: <AllInboxRoundedIcon style={{ fontSize: '18px' }} />,
+              active: router.route.startsWith(RouteNames.EMAIL_INBOXES),
+              // isExperimental: true,
+              isNew: false,
+            },
+            {
+              label: 'Contacts',
+              route: RouteNames.CONTACTS,
+              icon: <RecentActorsIcon style={{ fontSize: '18px' }} />,
+              active: router.route.startsWith(RouteNames.CONTACTS),
+              isNew: false,
+            },
+          ]
+        : []),
+      ...(product === 'cs'
+        ? [
+            {
+              label: 'Inbox',
+              route: RouteNames.LOGS,
+              icon: (
+                <Badge
+                  badgeContent={countUnreadQuery?.data}
+                  size="sm"
+                  color="danger"
+                  invisible={
+                    !countUnreadQuery?.data || countUnreadQuery?.data <= 0
+                  }
+                >
+                  <InboxRoundedIcon style={{ fontSize: '18px' }} />
+                </Badge>
+              ),
+              active: router.route === RouteNames.LOGS,
+              isNew: false,
+            },
+            {
+              label: 'Agents',
+              route: RouteNames.AGENTS,
+              icon: <SmartToyRoundedIcon style={{ fontSize: '18px' }} />,
+              active: router.route.startsWith(RouteNames.AGENTS),
+              isExperimental: false,
+              isNew: false,
+            },
+          ]
+        : []),
+      ...(product === 'chat'
+        ? [
+            {
+              label: 'Chat',
+              route: RouteNames.CHAT,
+              icon: <ChatRoundedIcon style={{ fontSize: '18px' }} />,
+              active: router.route === RouteNames.CHAT,
+              isExperimental: false,
+              isNew: false,
+            },
+            {
+              label: 'Datastores',
+              route: RouteNames.DATASTORES,
+              icon: <StorageRoundedIcon style={{ fontSize: '18px' }} />,
+              active: router.route.startsWith(RouteNames.DATASTORES),
+              isNew: false,
+            },
+          ]
+        : []),
+    ];
+  }, [router.route, countUnreadQuery?.data, product]);
+
+  const settingLinks = React.useMemo(() => {
+    return [
+      // {
+      //   label: 'Apps',
+      //   route: RouteNames.APPS,
+      //   icon: <AutoFixHighRoundedIcon fontSize="small" />,
+      //   active: router.route === RouteNames.APPS,
+      // },
+      {
+        label: 'Settings',
+        route: RouteNames.SETTINGS,
+        icon: <ManageAccountsRoundedIcon style={{ fontSize: '18px' }} />,
+        active: router.route.startsWith(RouteNames.SETTINGS),
+        isExperimental: false,
+        isNew: false,
+      },
+    ];
+  }, [router.route]);
+
+  const docLinks = React.useMemo(() => {
+    return [
+      {
+        label: 'Documentation',
+        route: 'https://docs.chaindesk.ai/',
+        icon: <ApiRoundedIcon style={{ fontSize: '18px' }} />,
+        target: 'blank',
+        isExperimental: false,
+        isNew: false,
+      },
+      // {
+      //   label: 'Help Center',
+      //   route: 'https://chaindesk.ai/help',
+      //   icon: <HelpRoundedIcon fontSize="small" />,
+      //   target: 'blank',
+      //   isExperimental: false,
+      //   isNew: false,
+      // },
+    ];
+  }, [router.route]);
+
+  const { publicRuntimeConfig } = getConfig();
+  const isMaintenance = !!getStatusQuery?.data?.isMaintenance;
+
+  const SystemStatusIndicator = getStatusQuery?.data?.status ? (
+    <Link
+      href={'https://status.chaindesk.ai/'}
+      target={'_blank'}
+      className={!open ? 'fixed bottom-2' : ''}
+    >
+      <Chip
+        color={
+          (
+            {
+              [AppStatus.OK]: 'success',
+              [AppStatus.WARNING]: 'warning',
+              [AppStatus.KO]: 'danger',
+            } as Record<AppStatus, ColorPaletteProp>
+          )[getStatusQuery?.data?.status]
+        }
+        variant="soft"
+        sx={{ cursor: 'pointer' }}
+        endDecorator={open ? <ArrowForwardRoundedIcon /> : null}
+      >
+        <Stack direction="row" alignItems={'center'} gap={1}>
+          <Box
+            sx={{
+              width: '10px',
+              height: '10px',
+              borderRadius: '99px',
+              // bgcolor: isStatusOK ? 'success.300' : 'danger.500',
+              ...(getStatusQuery?.data?.status === AppStatus.OK && {
+                bgcolor: 'success.300',
+              }),
+              ...(getStatusQuery?.data?.status === AppStatus.KO && {
+                bgcolor: 'danger.500',
+              }),
+              ...(getStatusQuery?.data?.status === AppStatus.WARNING && {
+                bgcolor: 'warning.500',
+              }),
+            }}
+          />
+          {open && <Typography level="body-sm">system status</Typography>}
+        </Stack>
+      </Chip>
+    </Link>
+  ) : (
+    <div />
+  );
+
+  React.useEffect(() => {
+    if (typeof window !== 'undefined') {
+      if (
+        window.location.hostname === 'app.databerry.ai' ||
+        window.location.hostname === 'www.chaindesk.ai' ||
+        window.location.hostname === 'chaindesk.ai'
+      ) {
+        window.location.href =
+          'https://app.chaindesk.ai' + window.location.pathname;
+      }
+    }
+  }, []);
+
+  React.useEffect(() => {
+    if (
+      getStatusQuery?.data?.status &&
+      getStatusQuery?.data?.status !== AppStatus.OK
+    ) {
+      toast.error(
+        "We're experiencing some issues. Please try again later. Sorry for the inconvenience!",
+        {
+          duration: 100000,
+          id: 'status-error',
+        }
+      );
+    }
+  }, [getStatusQuery?.data?.status]);
+
   // const showPromoBanner =
   //   status !== 'loading' && !session?.organization?.isPremium;
   const showPromoBanner = false;
@@ -83,7 +360,14 @@ export default function Layout(props: Props) {
           className={mounted ? mode : ''}
         >
           <Box sx={{ height: '100%', overflowY: 'auto' }}>
-            <Navigation />
+            <ExpandedNavigation
+              product={product}
+              appLinks={appLinks}
+              docLinks={docLinks as any}
+              settingLinks={settingLinks}
+              publicRuntimeConfig={publicRuntimeConfig}
+              status={getStatusQuery?.data?.status}
+            />
           </Box>
         </SideDrawer>
       )}
@@ -165,7 +449,33 @@ export default function Layout(props: Props) {
           overflow: 'hidden',
         }}
       >
-        <Navigation />
+        <IconButton
+          variant="outlined"
+          size="sm"
+          onClick={() => {
+            setOpen(true);
+            setDrawerOpen(true);
+          }}
+          sx={{
+            display: { sm: 'none' },
+            marginTop: '0.6rem',
+            left: -10,
+            zIndex: 999,
+            position: 'fixed',
+          }}
+        >
+          <ArrowForwardIosIcon />
+        </IconButton>
+        <Navigation
+          latestVersion={getStatusQuery?.data?.latestVersion}
+          product={product}
+          appLinks={appLinks}
+          docLinks={docLinks as any}
+          settingLinks={settingLinks}
+          publicRuntimeConfig={publicRuntimeConfig}
+          status={getStatusQuery?.data?.status}
+          isMaintenance={isMaintenance}
+        />
 
         <Main
           sx={{
@@ -173,6 +483,7 @@ export default function Layout(props: Props) {
             minHeight: `calc(100dvh - ${promoBannerHeight}px)`,
             maxHeight: `calc(100dvh - ${promoBannerHeight}px)`,
             height: '100dvh',
+            overflowY: 'auto',
             width: '100%',
 
             ...props.mainSxProps,
